+++ config/config.go
@@ -46,6 +46,10 @@ type Config struct {
 	Metrics *MetricsConfig `group:"metrics" namespace:"metrics"`
 
 	BabylonConfig *BBNConfig `group:"babylon" namespace:"babylon"`
+
+	RemoteSignerEnabled bool `long:"remote-signer-enabled" description:"if true, covenant will use the remote signer to sign transactions"`
+
+	RemoteSigner *RemoteSignerCfg `group:"remotesigner" namespace:"remotesigner"`
 }
 
 // LoadConfig initializes and parses the config using a config file and command
@@ -128,15 +132,18 @@ func DefaultConfigWithHomePath(homePath string) Config {
 	bbnCfg.Key = defaultCovenantKeyName
 	bbnCfg.KeyDirectory = homePath
 	metricsCfg := DefaultMetricsConfig()
+	remoteSignerCfg := DefaultRemoteSignerConfig()
 	cfg := Config{
-		LogLevel:        defaultLogLevel,
-		QueryInterval:   defaultQueryInterval,
-		DelegationLimit: defaultDelegationLimit,
-		SigsBatchSize:   defaultSigsBatchSize,
-		BitcoinNetwork:  defaultBitcoinNetwork,
-		BTCNetParams:    defaultBTCNetParams,
-		Metrics:         &metricsCfg,
-		BabylonConfig:   &bbnCfg,
+		LogLevel:            defaultLogLevel,
+		QueryInterval:       defaultQueryInterval,
+		DelegationLimit:     defaultDelegationLimit,
+		SigsBatchSize:       defaultSigsBatchSize,
+		BitcoinNetwork:      defaultBitcoinNetwork,
+		BTCNetParams:        defaultBTCNetParams,
+		Metrics:             &metricsCfg,
+		BabylonConfig:       &bbnCfg,
+		RemoteSignerEnabled: false,
+		RemoteSigner:        &remoteSignerCfg,
 	}
 
 	if err := cfg.Validate(); err != nil {